Split layout tests
