+++ weasyprint/tests/test_layout/__init__.py
@@ -1,0 +1,11 @@
+"""
+    weasyprint.tests.layout
+    -----------------------
+
+    Tests for layout, ie. positioning and dimensioning of boxes,
+    line breaks, page breaks.
+
+    :copyright: Copyright 2011-2018 Simon Sapin and contributors, see AUTHORS.
+    :license: BSD, see LICENSE for details.
+
+"""
